+++ projects/ix-libs/schematics/ng-add/schema.ts
@@ -1,0 +1,7 @@
+export interface Schema {
+    project?: string;
+    docsBaseRoute: string;
+    assetsRoot: string;
+    titleSuffix: string;
+    skipRoute: boolean;
+}